requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
asyncio==3.4.3
attrs==25.1.0
azure-ai-vision-imageanalysis==1.0.0
azure-core==1.32.0
backoff==2.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
cryptography==44.0.1
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
duckduckgo_search==7.3.2
emoji==2.14.1
eval_type_backport==0.2.2
filelock==3.17.0
filetype==1.2.0
fire==0.7.0
Flask==3.1.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.2.0
get_project_root==0.2
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-vision==3.10.0
googleapis-common-protos==1.67.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
langchain==0.3.18
langchain-anthropic==0.3.7
langchain-community==0.3.17
langchain-core==0.3.34
langchain-openai==0.3.5
langchain-text-splitters==0.3.6
langdetect==1.0.9
langgraph==0.2.71
langgraph-checkpoint==2.0.12
langgraph-sdk==0.1.51
langsmith==0.3.8
lxml==5.3.1
markdown-it-py==3.0.0
markdownify==0.14.1
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
olefile==0.47
onnxruntime==1.20.1
openai==1.61.1
opencv-python==4.11.0.86
orjson==3.10.15
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
playwright==1.50.0
primp==0.12.1
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyee==12.1.1
Pygments==2.19.1
pymongo==4.11.1
pypdf==5.3.0
pyperclip==1.9.0
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2025.2.8
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2025.1
PyYAML==6.0.2
randomname==0.2.1
RapidFuzz==3.12.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
safetensors==0.5.2
selenium==4.28.1
sentencepiece==0.2.0
shapely==2.0.7
six==1.17.0
smolagents==1.8.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.38
sympy==1.13.3
tarsier==0.8.2
tenacity==9.0.0
termcolor==2.5.0
tiktoken==0.8.0
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.48.3
trio==0.28.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
unstructured==0.16.20
unstructured-client==0.29.0
urllib3==2.3.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zstandard==0.23.0
